moved files, add permissions check
